Fix failing specs (#44)
* Fix fact-related telegraf::input specs

These failures appear to have been caused by additional fact requirements
added since the specs were first written. Additionally, the my_snmp
check was failing due to a change that causes integers to be written to
the config as bare numerals, not as strings.

* Bump Travis ruby version

Required for newer versions of beaker.