Allow input/output sections to be added as 2-element array
Plugins in Telegraf allow for wildly different configuration
styles, making use of the TOML configuration language features.

One of these features is the ability to specify arrays of tables
and even nesting them. This is for example used by the `inputs.snmp`
plugin to specify multiple SNMP queries and their associated
special casing of certain SNMP values. For example:

```toml
[[inputs.snmp]]
  agents = [ "sw0-0.example.com:161", "sw0-1.example.com:161", "sw0-2.example.com:161" ]
  version = 2
  interval = "60s"

[[inputs.snmp.table]]
  ## auto populate table's fields using the MIB
  oid = "IF-MIB::ifTable"

[[inputs.snmp.table.field]]
  oid = "IF-MIB::ifType"
  is_tag = true

[[inputs.snmp.table.field]]
  oid = "IF-MIB::ifDescr"
  is_tag = true

[[inputs.snmp.table.field]]
  oid = "IF-MIB::ifPhysAddress"
  is_tag = true

[[inputs.snmp.table]]
  ## auto populate table's fields using the MIB
  oid = "IF-MIB::ifXTable"

[[inputs.snmp.table.field]]
  oid = "IF-MIB::ifName"
  is_tag = true

[[inputs.snmp.table.field]]
  oid = "IF-MIB::ifAlias"
  is_tag = true
```

Having only support for one instance of a section name and also
sorting these makes it impossible to use Telegraf plugins like
the SNMP one correctly.

Therefore this commit introduces the ability to alternatively
specify the sections as a 2-element array in the format of
[<section name>, { option => value, other_option => value }].