Default 'options' of input result in incorrect config
If you were to specify an input without any configuration using `telegraf::input`, the resulting telegraf config file has an incorrect format:

```
  telegraf::input { 'mem':
    plugin_type => 'mem',
  }
```

results in the following `mem.conf` being created:

```
[inputs]
mem = []
```

This is not a valid telegraf config and it will fail on this.

I suspect this is because the `stdlib::to_toml({ 'inputs'=> { $plugin_type=> $options } })` just creates a key-value pair if `$options` does not contain a hash. It is easily remedied by passing in `options => [{}]`, however this is not really intuitive to figure out, hence I'd suggest having a functional default value.

Additionally it wasn't really clear to me why this needed to be an Array
to begin with, since a Hash seems more fitting, however given that this
was already used in [`params.pp`](https://github.com/voxpupuli/puppet-telegraf/blob/master/manifests/params.pp#L142-L147) like this as well, I opted to leave it like this for now.